{{cookiecutter.project_name}}/templates/multiner/requirements.txt
transformers==4.34.0
torch==2.1.0
aina-gradio-theme==1.0
gradio=={{cookiecutter.gradio_version}}
python-dotenv==1.0.0
